VZ-7666 part 2 - Move VMC controller to cluster-operator
