+++ cluster-operator/controllers/vmc/sync_registration_secret.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2021, 2022, Oracle and/or its affiliates.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 
-package clusters
+package vmc
 
 import (
 	"context"
